refactor(backend): relocate `Email` struct to `utils`
Move the `Email` struct to the universal `utils/` directory to align
with our backend structure. Page-specific utilities, like custom email
templates, remain in `services/`.
