+++ backend/src/services/auth.rs
@@ -7,12 +7,13 @@ use crate::routes::auth::AuthError::{
 use crate::routes::auth::RegisterModel;
 use crate::routes::auth::reset_password::ResetPasswordModel;
 use crate::services::emails;
-use crate::services::emails::{Email, EmailConfirmation};
+use crate::services::emails::EmailConfirmation;
 use crate::services::env::EnvConfig;
 use crate::utils::app_state;
 use crate::utils::cookies::{
     ACCESS_TOKEN_COOKIE, REFRESH_TOKEN_COOKIE, create_cookie, reset_cookie,
 };
+use crate::utils::email::Email;
 use crate::utils::error::Error;
 use crate::utils::error::Error::HashPasswordFailed;
 use crate::utils::jwt::encode_jwt;
